+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -8,6 +8,7 @@ import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 import net.runelite.client.config.ConfigSection;
+import net.runelite.client.config.Range;
 
 @ConfigGroup("dinkplugin")
 public interface DinkPluginConfig extends Config {
@@ -103,6 +104,13 @@ public interface DinkPluginConfig extends Config {
         position = 110
     )
     String diarySection = "Achievement Diary";
+
+    @ConfigSection(
+        name = "BA Gambles",
+        description = "Settings for notifying when you gamble at Barbarian Assault",
+        position = 120
+    )
+    String gambleSection = "BA Gambles";
 
     @ConfigSection(
         name = "Advanced",
@@ -340,6 +348,17 @@ public interface DinkPluginConfig extends Config {
         section = webhookSection
     )
     default String diaryWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "gambleWebhook",
+        name = "BA Gamble Webhook Override",
+        description = "If non-empty, BA gamble messages are sent to this URL, instead of the primary URL",
+        position = -7,
+        section = webhookSection
+    )
+    default String gambleWebhook() {
         return "";
     }
 
@@ -1051,4 +1070,76 @@ public interface DinkPluginConfig extends Config {
         return "%USERNAME% has completed the %DIFFICULTY% %AREA% Achievement Diary, for a total of %TOTAL% diaries completed";
     }
 
+    @ConfigItem(
+        keyName = "gambleEnabled",
+        name = "Enable BA Gamble",
+        description = "Enable notifications for Barbarian Assault high level gambles",
+        position = 120,
+        section = gambleSection
+    )
+    default boolean notifyGamble() {
+        return false;
+    }
+
+    @ConfigItem(
+        keyName = "gambleSendImage",
+        name = "Send Image",
+        description = "Send image with the notification",
+        position = 121,
+        section = gambleSection
+    )
+    default boolean gambleSendImage() {
+        return true;
+    }
+
+    @ConfigItem(
+        keyName = "gambleInterval",
+        name = "Notify Interval",
+        description = "Interval between when a notification should be sent",
+        position = 122,
+        section = gambleSection
+    )
+    @Range(min = 1)
+    default int gambleInterval() {
+        return 100;
+    }
+
+    @ConfigItem(
+        keyName = "gambleRareLoot",
+        name = "Always notify for rare loot",
+        description = "Always send a notification upon receiving a dragon chainbody or med helm from a gamble",
+        position = 123,
+        section = gambleSection
+    )
+    default boolean gambleRareLoot() {
+        return true;
+    }
+
+    @ConfigItem(
+        keyName = "gambleNotifMessage",
+        name = "Notification Message",
+        description = "The message to be sent through the webhook every gamble interval.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %COUNT% to insert the gamble count<br/>" +
+            "Use %LOOT% to insert the loot",
+        position = 124,
+        section = gambleSection
+    )
+    default String gambleNotifyMessage() {
+        return "%USERNAME% has reached %COUNT% high gambles";
+    }
+
+    @ConfigItem(
+        keyName = "gambleRareNotifMessage",
+        name = "Rare Loot Notification Message",
+        description = "The message to be sent through the webhook for rare loot.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %COUNT% to insert the gamble count<br/>" +
+            "Use %LOOT% to insert the loot",
+        position = 125,
+        section = gambleSection
+    )
+    default String gambleRareNotifyMessage() {
+        return "%USERNAME% has received rare loot at gamble count %COUNT%: \n\n%LOOT%";
+    }
 }